+++ pronamic-pay-mollie.php
@@ -4,8 +4,9 @@
  * Plugin URI: https://www.pronamic.eu/plugins/pronamic-pay-mollie/
  * Description: Extend the Pronamic Pay plugin with the Mollie gateway to receive payments through a variety of WordPress plugins.
  *
- * Version: 4.5.0
- * Requires at least: 4.7
+ * Version: 4.7.6
+ * Requires at least: 5.9
+ * Requires PHP: 7.4
  *
  * Author: Pronamic
  * Author URI: https://www.pronamic.eu/
@@ -15,12 +16,13 @@
  *
  * License: GPL-3.0-or-later
  *
+ * Requires Plugins: pronamic-ideal
  * Depends: wp-pay/core
  *
  * GitHub URI: https://github.com/pronamic/wp-pronamic-pay-mollie
  *
  * @author    Pronamic <info@pronamic.eu>
- * @copyright 2005-2022 Pronamic
+ * @copyright 2005-2023 Pronamic
  * @license   GPL-3.0-or-later
  * @package   Pronamic\WordPress\Pay\Gateways\Mollie
  */