+++ tests/bootstrap.php
@@ -3,7 +3,7 @@
  * Bootstrap tests
  *
  * @author    Pronamic <info@pronamic.eu>
- * @copyright 2005-2022 Pronamic
+ * @copyright 2005-2023 Pronamic
  * @license   GPL-3.0-or-later
  * @package   Pronamic\WordPress\Pay\Core
  */
@@ -22,16 +22,16 @@ require_once __DIR__ . '/../vendor/autoload.php';
  * Psalm.
  */
 if ( defined( 'PSALM_VERSION' ) ) {
-    return;
+	return;
 }
 
 /**
  * Plugin.
  */
 $plugin = \Pronamic\WordPress\Pay\Plugin::instance(
-    array(
-        'action_scheduler' => __DIR__ . '/../vendor/woocommerce/action-scheduler/action-scheduler.php',
-    )
+	array(
+		'action_scheduler' => __DIR__ . '/../vendor/woocommerce/action-scheduler/action-scheduler.php',
+	)
 );
 
 $plugin->plugins_loaded();